+++ emacs/ocamlformat.el
@@ -294,6 +294,112 @@ is nil."
     (delete-file bufferfile)
     (delete-file outputfile)))
 
+(defun ocamlformat-args (name start-line end-line)
+  (let*
+      ((margin-args
+        (cond
+         ((equal ocamlformat-margin-mode 'window)
+          (list "--margin" (number-to-string (window-body-width))))
+         ((equal ocamlformat-margin-mode 'fill)
+          (list "--margin" (number-to-string fill-column)))
+         (t
+          '())))
+       (enable-args
+        (cond
+         ((equal ocamlformat-enable 'disable)
+          (list "--disable"))
+         ((equal ocamlformat-enable 'enable-outside-detected-project)
+          (list "--enable-outside-detected-project"))
+         (t
+          '())))
+       (extension-args
+        (cond
+         ((eq ocamlformat-file-kind 'implementation)
+          (list "--impl"))
+         ((eq ocamlformat-file-kind 'interface)
+          (list "--intf")))))
+    (append margin-args enable-args extension-args
+            (list
+             "-"
+             "--name" name
+             "--numeric" (format "%d-%d" start-line end-line)))))
+
+(defun ocamlformat-region (start end)
+  (interactive "r")
+  (let*
+      ((ext (file-name-extension buffer-file-name t))
+       (bufferfile (file-truename (make-temp-file "ocamlformat" nil ext)))
+       (errorfile (file-truename (make-temp-file "ocamlformat" nil ext)))
+       (errbuf
+        (cond
+         ((eq ocamlformat-show-errors 'buffer)
+          (get-buffer-create "*compilation*"))
+         ((eq ocamlformat-show-errors 'echo)
+          (get-buffer-create "*OCamlFormat stderr*"))))
+       (start-line (line-number-at-pos start))
+       (end-line (line-number-at-pos end))
+       (indents-str
+        (with-output-to-string
+          (if (/= 0
+                (apply 'call-process-region
+                       (point-min) (point-max) ocamlformat-command nil
+                       (list standard-output errorfile) nil
+                       (ocamlformat-args buffer-file-name start-line end-line)))
+              (progn
+                (if errbuf
+                  (progn
+                    (with-current-buffer errbuf
+                      (setq buffer-read-only nil)
+                      (erase-buffer))
+                    (ocamlformat--process-errors
+                     (buffer-file-name) bufferfile errorfile errbuf)))
+                (message "Could not apply ocamlformat")))))
+       (indents (mapcar 'string-to-number (split-string indents-str))))
+    (save-excursion
+      (goto-char start)
+      (mapcar
+       #'(lambda (indent) (indent-line-to indent) (forward-line))
+       indents))
+    (delete-file errorfile)
+    (delete-file bufferfile)))
+
+(defun ocamlformat-line ()
+  (interactive nil)
+  (ocamlformat-region (point) (point)))
+
+;;;###autoload
+(defun ocamlformat-setup-indent ()
+  (interactive nil)
+  (set (make-local-variable 'indent-line-function) #'ocamlformat-line)
+  (set (make-local-variable 'indent-region-function) #'ocamlformat-region))
+
+;;;###autoload
+(defun ocamlformat-caml-mode-setup ()
+  (ocamlformat-setup-indent)
+  (local-unset-key "\t"))  ;; caml-mode rebinds TAB !
+
+(defun custom_newline-and-indent (&optional arg)
+  "Insert a newline, then indent according to `ocamlformat-line'.
+With ARG, perform this action that many times."
+  (interactive "*p")
+  (delete-horizontal-space t)
+  (unless arg
+    (setq arg 1))
+  (dotimes (_ arg)
+    (newline nil t)
+    (insert "x")
+    (ocamlformat-line)
+    (delete-backward-char 1 nil)))
+
+(defun set-newline-and-indent ()
+  (local-set-key (kbd "RET") 'custom_newline-and-indent))
+
+(add-hook 'tuareg-mode-hook 'ocamlformat-setup-indent t)
+(add-hook 'tuareg-mode-hook 'set-newline-and-indent)
+
+(add-hook 'caml-mode-hook 'ocamlformat-caml-mode-setup  t)
+(add-hook 'caml-mode-hook 'set-newline-and-indent)
+
 (provide 'ocamlformat)
 
 ;;; ocamlformat.el ends here