+++ lib/Source.mli
@@ -98,3 +98,5 @@ val is_quoted_string : t -> Location.t -> bool
 
 val loc_of_first_token_at :
   t -> Location.t -> Parser.token -> Location.t option
+
+val find_first_token_on_line : t -> int -> (Parser.token * Location.t) option